backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.20.0
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
